master: Bind webhook to listen on both IPv6 and IPv4
The webhook script defaults to listening on "0.0.0.0", which creates
an IPv4-only socket.  Fix by explicitly setting the bind address to
"*" in order to also listen on IPv6.

Note that we cannot use a bind address of "::" since values matching
the regexp /^:/ are special-cased by the webhook.yaml.erb template.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>