autosecret: Add autosecret::sha256 and autosecret::sha1 functions
Add ability to automatically generate secrets via e.g.

  autosecret::sha256('myapp', 'database')
  autosecret::sha256('myapp', 'admin')

The function accepts an arbitrary list of nonces that are used to
uniquely identify a password.  The actual password content will be
generated as a SHA-256 hash of the nonces plus a base secret.

The base secret will be retrieved from Hiera if present via
lookup('autosecret::base').  If no secret is found, then a random
secret will be generated and persisted into the Puppet server's
private information directory.

Multiple manifests may use the same password by simply using the same
list of nonces.  For example:

  autosecret::sha256('myapp', 'database')

may be used in one manifest to create a database user, and in a second
manifest to provide the database credentials to a client.

The net effect is that a single base secret may be used to generate an
arbitrary number of passwords.  Knowledge of any one password does not
provide access to any other password or to the base secret itself.  If
the base secret is stored encrypted (e.g. using EYAML), then all
passwords are effectively stored encrypted.

This code may eventually be published as a standalone Puppet module.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>