base: Avoid conflicts between ud::packages and explicitly managed packages
Puppet does not allow a resource to be declared more than once.  Allow
for duplicates between packages mentioned in ud::packages and packages
explicitly managed by other manifests, to avoid irritating scenarios
in which a package has to be removed from a project's common base list
because it happens to be used by a manifest applied to a single node.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>