apache: Use conditional checks for certificate existence
Conditionally check for the existence of the certificate using Apache
directives, using the default localhost certificate as a temporary
substitute if the real certificate is not yet available.

This allows Apache to start up prior to the certificate having been
issued, without requiring the Apache manifest to define resources for
the LetsEncrpyt certificate files and directories.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>