master: Generate SSH deploy key prior to repo configuration
There is currently a circular dependency: the Puppet master manifest
will generate a deploy key only if the repo has been configured, but
the unipart-puppet-setup tool requires the deploy key in order to
configure the repo.

Fix by generating the deploy key unconditionally, since doing so
requires no knowledge of which repo is to be used.

Debugged-by: Alvise Pellegrini <alvise.pellegrini@unipart.io>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>