master: Allow for local development branches
The existing mechanism does not modify $basemodulepath and so modules
from the development branch will always be ignored in favour of the
production branch.

Interpolating $environment into $basemodulepath is forbidden and so
cannot be used to effect the development model as currently
documented.

Switch to a model in which the Puppet master is configured to allow
for a single development branch named after the project in which it is
running.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>