postgresql: Enable TLS using LetsEncrypt certificate
Enable TLS, include the use of sslmode=verify-full in the generated
database URL, and allow TLS connections from anywhere (subject to
external security group rules).

The default configuration for the PostgreSQL client libraries is not
to bother verifying the server certificates.  It can be forced to do
so by setting the connection parameter sslmode=verify-full.  It fails
to use the system CA certificates as a root of trust.  Create a
symlink for all users managed by ud::user (on systems with or without
ud::server::postgresql) to minimise annoyance.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>