postgresql: Force array result from plusignment operator
Puppet's plusignment operator (+>) does not take into account the
default value or declared type of a parameter, and will instead use a
flawed heuristic of guessing between an array and a scalar value based
on the number of values assigned.

Fix by explicitly assigning an array containing a single element.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>