php: Use Remi RPM repository only for CentOS
The Remi modular repository for Fedora includes broken dependencies.
Since Fedora provides many more PHP packages than CentOS anyway,
assume for now that we do not need to enable a Remi module on Fedora.

Leave the repository configuration in place (but disabled) for the
inevitable moment when some PHP application requires a package that is
not provided by base Fedora.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>