apache: Disable TLSv1.3
TLSv1.3 drops support for renegotiation.  This breaks the ability to
renegotiate to select a client certificate, and hence breaks the use
of "SSLVerifyClient optional".

The replacement for renegotiation seems to be "post-handshake
authentication", but this is not yet supported by either Firefox or
Chrome.

Disable TLSv1.3 until post-handshake authentication is more widely
supported.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>